requirements.txt
django                     
django-tastypie      
mysql-connector-python    
mysql-python             
mysqlclient           
openssl                          
wheel     
requests  
pymongo 
